+++ src/test/converter2/issues/gh2466.ts
@@ -1,0 +1,15 @@
+// Conversion order test for @link
+
+export interface One extends Two {}
+
+/** {@link method1} */
+export interface Two {
+    method1(): string;
+}
+
+/** {@link method2} */
+export interface Three {
+    method2(): string;
+}
+
+export interface Four extends Three {}