+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 -   Fix crash when converting some complicated union/intersection types, #2451.
 -   Navigation triangle markers should no longer display on a separate line with some font settings, #2457.
 -   `@group` and `@category` organization is now applied later to allow inherited comments to create groups/categories, #2459.
+-   Conversion order should no longer affect link resolution for classes with properties whose type does not rely on `this`, #2466.
 -   Keyword syntax highlighting introduced in 0.25.4 was not always applied to keywords.
 -   If all members in a group are hidden from the page, the group will be hidden in the page index on page load.